+++ bin/commands/attackpayloadlogging.js
@@ -1,11 +1,10 @@
-let ChallengeInterceptionRules = require('../../src/challengeinterceptionrules')
-  .challengeinterceptionrules;
+let AdvancedSettings = require('../../src/advancedsettings').advancedsettings;
 let out = require('./lib/out');
 
-class ChallengeInterceptionRulesCommand {
+class AttackPayloadLoggingCommand {
   constructor() {
-    this.flags = 'challenge-interception-rules';
-    this.desc = '(Deprecated) Display contents of challenge interception rules.';
+    this.flags = 'attack-payload-logging';
+    this.desc = 'Display the Attack Payload Logging settings.';
     this.setup = this.setup.bind(this);
     this.run = this.run.bind(this);
   }
@@ -22,12 +21,18 @@ class ChallengeInterceptionRulesCommand {
           "Version Number. It can also take the values 'PROD' or 'PRODUCTION' or 'STAGING'. If not provided, latest version is assumed.",
         group: 'Optional:',
         required: false
+      })
+      .string('--policy <id>', {
+        desc:
+          'Policy ID. If provided, returns policy-level settings. If not provided, returns config-level settings.',
+        group: 'Optional:',
+        required: false
       });
   }
 
   run(options) {
     out.print({
-      promise: new ChallengeInterceptionRules(options).getChallengeInterceptionRules(),
+      promise: new AdvancedSettings(options).getAttackPayloadLogging(),
       args: options,
       success: (args, data) => {
         return JSON.stringify(data);
@@ -36,4 +41,4 @@ class ChallengeInterceptionRulesCommand {
   }
 }
 
-module.exports = new ChallengeInterceptionRulesCommand();
+module.exports = new AttackPayloadLoggingCommand();